refactor(apollo_integration_tests): only EXTRACT, to run flow tests sequentially
Changes:
ONLY MOVE, no logic changes. Break up the rstest
parameterization into two separate files inside `tests/`, which
corresponds to two `cargo`-style integration tests.
Also moved a bunch of _single-use_ functions from utils.rs into the
corresponding test file.

Motivation:
* https://doc.rust-lang.org/rust-by-example/testing/integration_testing.html#integration-testing.
Every file in `tests/` is run as a separate binary by cargo, without
parallelization.
However, multiple tests in the same file in `tests/` are still parallelized by cargo,
which is what `rstest` parametrization generates.

* This is the best way to run flow tests (and also integration) for
a bunch of reasons: we don't want stuff in one test contaminating the
other, we want to use fixed (configurable) ports for things (there is no
requirement to do otherwise and dynamic ports adds complexity to the
setup).

* Clarifying note: in the future, we want the _integration tests_ to become tests
  inside `tests/` inside `apollo_node`, for somewhat similar reasons,
  plus they will spare us from using the bash script instead of `cargo test`,
  but this is unrelated to this commit.
